gpu: nvgpu: clk: fix MISRA 10.3 issue in clk_prog
MISRA Rule 10.3 prohibits implicit assignment of an object from a
different size type. This fixes a MISRA 10.3 violation for assigning a
u16 to a u8 in clk_prog.c.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2001227
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/clk/clk_prog.c
@@ -1410,8 +1410,9 @@ static int vfflatten_prog_1x_master(struct gk20a *g,
 			freq_step_size_mhz =
 				p1xmaster->super.source_data.pll.freq_step_size_mhz;
 			step_count = (freq_step_size_mhz == 0U) ? 0U :
-				(p1xmaster->super.freq_max_mhz - *pfreqmaxlastmhz - 1U) /
-							freq_step_size_mhz;
+					(u8)(p1xmaster->super.freq_max_mhz -
+						*pfreqmaxlastmhz - 1U) /
+					freq_step_size_mhz;
 			/* Intentional fall-through.*/
 
 		case CTRL_CLK_PROG_1X_SOURCE_ONE_SOURCE:
